+++ src/middleware/index.js
@@ -17,7 +17,6 @@ import filter from './filter';
 import pagination from './pagination';
 import search from './search';
 import sort from './sort';
-import validations from './validations';
 
 const auth = new Auth();
 const check = new Check();
@@ -30,5 +29,4 @@ export default {
   pagination,
   search,
   sort,
-  validations
 };
