+++ src/routes/occupations.js
@@ -8,8 +8,8 @@
 import { Router } from 'express';
 import controllers from '../controllers';
 import middlewares from '../middleware';
-import validations from '../middleware/validations';
 import sort from '../middleware/sort';
+import { occupations as validations } from '../validations';
 
 const { auth, check } = middlewares;
 
@@ -21,7 +21,7 @@ routes.use(auth.authenticateUser, check.currentUserIsAdmin);
 
 routes.post(
   '/',
-  validations.occupations.addOccupation,
+  validations.addOccupation,
   controllers.occupations.addOccupation
 );
