+++ src/validations/opportunities.js
@@ -9,7 +9,7 @@
 import Validator from 'validatorjs';
 
 import { getErrors } from './';
-import models from '../../models';
+import models from '../models';
 
 const opportunityUploadRules = {
   title: 'required|string|max:36',
