+++ src/validations/tags.js
@@ -1,6 +1,6 @@
 import Validator from 'validatorjs';
 import { getErrors } from '.';
-import models from '../../models';
+import models from '../models';
 
 const createMajorTagRules = {
   title: 'required|string'
